dev-ifc2tileset-agent: sanitise __init__.py files
